refactor folder structure done
