+++ composer.json
@@ -13,7 +13,7 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": "^7.1",
+    "php": "^7.4",
     "mustache/mustache": "^2.13.0",
     "psr/log": "^1.0",
     "matthiasmullie/minify": "^1.3",
