+++ view/react/material-dashboard-react-c6/.babelrc
@@ -5,7 +5,7 @@
     "transform-react-jsx",
     "transform-object-rest-spread",
     ["module-resolver", {
-      "root": ["./src"],
+      "root": ["./src"]
     }],
     ["import-rename", {"^(.*)\\.jsx$": "$1"}]
   ]
